split off helper methods into their own file, altered tests to be less strict with the file contents, added output test, removed unneeded import from alarm_clock, added docstrings
