+++ python/tests/alarm_clock/alarm_clock.yaml
@@ -1,6 +1,5 @@
 import:
   files:
-    - ./structures.yaml
     - ../alarm_clock/structures.yaml
 ---
 model:
